+++ src/auth/auth.module.ts
@@ -1,10 +1,11 @@
 import { AuthService } from './auth.service';
+import { GoogleOAuthService } from './google-o-auth.service';
 import { Module } from '@nestjs/common';
-import { OAuthService } from '@src/o-auth/o-auth.service';
+import { OAuthService } from './o-auth.service';
 import { PrismaService } from '@src/prisma/prisma.service';
 
 @Module({
-  providers: [AuthService, PrismaService, OAuthService],
+  providers: [AuthService, PrismaService, OAuthService, GoogleOAuthService],
   exports: [AuthService],
 })
 export class AuthModule {}
